Update meter_page : Seperated meter_info widget
